requirements.txt
matplotlib==3.8.2
networkx==3.2.1
langchain==0.3.12
langchain_ollama==0.2.1
spacy==3.7.2
pillow==10.2.0


